requirements.txt
# Google Photos API authentication
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
requests>=2.31.0

# SSH support
paramiko>=3.3.1
